requirements.txt
PyYAML==6.0.2
lm-eval[api]==0.4.9.1
psutil==7.0.0
pytest-timeout==2.4.0
pytest==8.4.1
requests==2.32.5
